// src/domain/enums.ts
export enum ShippingType {
  Delivery = "DELIVERY",
  Collect = "COLLECT",
}

export enum ShippingMethod {
  Scheduled = "SCHEDULED",
  SameDay = "SAME_DAY",
  NextDay = "NEXT_DAY",
  DateRange = "DATE_RANGE",
}

export enum ShippingSubType {
  InStore = "IN_STORE",
  Local = "LOCAL",
}

export enum ProductClassification {
  Regular = "REGULAR",
  Bundled = "BUNDLED",
  Sampling = "SAMPLING",
  Services = "SERVICES",
  Intangible = "INTANGIBLE",
  Presales = "PRESALES",
}

export enum SellerType {
  Paris = "1P",
  Marketplace = "3P",
}